fix: move page style to layout
